pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Explicit package target — required because the project has both a top-level
# `models/` directory (generated artifacts, NOT a Python package) and
# `src/pyro_dcm/models/` (the real subpackage). Without this, hatchling's
# auto-discovery ships a wheel missing `src/pyro_dcm/models/`, which breaks
# `from pyro_dcm.models import …` in `src/pyro_dcm/__init__.py`.
[tool.hatch.build.targets.wheel]
packages = ["src/pyro_dcm"]

[project]
name = "pyro-dcm"
version = "0.1.0"
description = "Neural amortized Dynamic Causal Modeling with Pyro"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "torch>=2.0",
    "torchdiffeq",
    "pyro-ppl>=1.9",
    "scipy",
    "numpy",
    "zuko>=1.2",
]

[project.optional-dependencies]
benchmark = [
    "matplotlib",
    "tabulate",
]
dev = [
    "pytest",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
]
ignore = [
    "D100",  # missing docstring in public module
    "D104",  # missing docstring in public package
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "spm: marks tests requiring MATLAB + SPM12 (deselect with '-m \"not spm\"')",
    "tapas: marks tests requiring tapas rDCM MATLAB toolbox",
]
